+++ packages/ember-rdfa-editor/src/commands/lift-empty-block-checked.ts
@@ -1,7 +1,7 @@
 import { canSplit, specCanSplit } from '#root/commands/split-block-checked.ts';
 import type { Command, TextSelection } from 'prosemirror-state';
 import { liftTarget } from 'prosemirror-transform';
-import { isNone, unwrap } from '#root/utils/_private/option.ts';
+import { isNone, unwrap } from '#root/utils/option.ts';
 
 export const liftEmptyBlockChecked: Command = (state, dispatch) => {
   const { $cursor } = state.selection as TextSelection;