+++ packages/ember-rdfa-editor/src/core/say-data-factory/literal.ts
@@ -1,6 +1,6 @@
 import type * as RDF from '@rdfjs/types';
 import { SayNamedNode } from './named-node.ts';
-import type { Option } from '#root/utils/_private/option.ts';
+import type { Option } from '#root/utils/option.ts';
 import type { SayTerm } from './term.ts';
 
 /**