+++ packages/ember-rdfa-editor/src/helpers/limit-text.ts
@@ -1,0 +1,10 @@
+import { helper } from '@ember/component/helper';
+
+export function limitText([string]: string[]) {
+  if (string && string.length > 140) {
+    return string.slice(0, 140) + '...';
+  } else {
+    return string;
+  }
+}
+export default helper(limitText);